samples: nrf9160: modem_shell: remove GPS driver support
Removed support for the deprecated GPS driver.

Fixed LNA configuration so that the default configuration
can be overridden in prj.conf.

Updated LNA configuration to use the new AT command interface
and removed some references to the old interface.
