+++ samples/nrf9160/modem_shell/Kconfig
@@ -121,21 +121,6 @@ config MOSH_LINK_SETT_NORMAL_MODE_AT_CMD_STR_LEN
 endmenu
 endif #MOSH_LINK
 
-choice MOSH_GNSS_API
-	bool "GNSS API"
-	depends on MOSH_GNSS
-	default MOSH_GNSS_INTERFACE
-	help
-	  Used GNSS API type.
-
-config MOSH_GNSS_INTERFACE
-	bool "GNSS interface"
-
-config MOSH_GNSS_GPS_DRIVER
-	bool "GPS driver"
-
-endchoice
-
 choice MOSH_GNSS_ANTENNA
 	depends on MOSH_GNSS
 	default MOSH_GNSS_ANTENNA_ONBOARD
@@ -148,7 +133,7 @@ config MOSH_GNSS_ANTENNA_EXTERNAL
 endchoice
 
 config MOSH_AT_MAGPIO
-	string
+	string "AT%XMAGPIO command"
 	default "AT\%XMAGPIO=1,0,0,1,1,1574,1577" if BOARD_NRF9160DK_NRF9160_NS
 	default "AT\%XMAGPIO=1,1,1,7,1,746,803,2,698,748,2,1710,2200,3,824,894,4,880,960,5,791,849,7,1565,1586" if BOARD_THINGY91_NRF9160_NS
 	help
@@ -156,7 +141,7 @@ config MOSH_AT_MAGPIO
 	  empty if this command should not be sent.
 
 config MOSH_AT_COEX0
-	string
+	string "AT%XCOEX0 command"
 	default "AT\%XCOEX0=1,1,1565,1586" if (BOARD_NRF9160DK_NRF9160_NS || BOARD_THINGY91_NRF9160_NS) && MOSH_GNSS_ANTENNA_ONBOARD
 	default "AT\%XCOEX0" if (BOARD_NRF9160DK_NRF9160_NS || BOARD_THINGY91_NRF9160_NS) && MOSH_GNSS_ANTENNA_EXTERNAL
 	help
