+++ samples/nrf9160/modem_shell/prj.conf
@@ -138,16 +138,6 @@ CONFIG_MULTICELL_LOCATION_SERVICE_NRF_CLOUD=y
 #CONFIG_MULTICELL_LOCATION_POLTE_CUSTOMER_ID="customer id"
 #CONFIG_MULTICELL_LOCATION_POLTE_API_TOKEN="api token"
 
-# Configurations for the GPS driver, uncomment all to use GPS driver instead of the GNSS interface.
-#CONFIG_MOSH_GNSS_GPS_DRIVER=y
-#CONFIG_NRF9160_GPS=y
-#CONFIG_NRF9160_GPS_NMEA_GSV=y
-#CONFIG_NRF9160_GPS_NMEA_GSA=y
-#CONFIG_NRF9160_GPS_NMEA_GLL=y
-#CONFIG_NRF9160_GPS_NMEA_GGA=y
-#CONFIG_NRF9160_GPS_NMEA_RMC=y
-#CONFIG_NRF9160_GPS_HANDLE_MODEM_CONFIGURATION=n
-
 # Library for buttons and LEDs
 CONFIG_DK_LIBRARY=y
